# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  rpc.InscriptionNumber:
    properties:
      classic:
        type: integer
      jubilee:
        type: integer
    type: object
  rpc.OrdHookBlock:
    properties:
      block_identifier:
        $ref: '#/definitions/rpc.OrdHookBlockIdentifier'
      metadata:
        additionalProperties: true
        type: object
      parent_block_identifier:
        $ref: '#/definitions/rpc.OrdHookBlockIdentifier'
      timestamp:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/rpc.OrdHookTransaction'
        type: array
    type: object
  rpc.OrdHookBlockIdentifier:
    properties:
      hash:
        type: string
      index:
        type: integer
    type: object
  rpc.OrdHookChain:
    properties:
      is_streaming_blocks:
        type: boolean
      predicate:
        $ref: '#/definitions/rpc.OrdHookPredicate'
      uuid:
        type: string
    type: object
  rpc.OrdHookDestination:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  rpc.OrdHookEvent:
    properties:
      apply:
        items:
          $ref: '#/definitions/rpc.OrdHookBlock'
        type: array
      chainhook:
        $ref: '#/definitions/rpc.OrdHookChain'
      rollback:
        description: Assuming rollback has a similar structure
        items: {}
        type: array
    type: object
  rpc.OrdHookInscriptionRevealed:
    properties:
      content_bytes:
        type: string
      content_length:
        type: integer
      content_type:
        type: string
      curse_type:
        description: Assuming curse_type can be null, hence using a pointer
        type: string
      inscriber_address:
        type: string
      inscription_fee:
        type: integer
      inscription_id:
        type: string
      inscription_input_index:
        type: integer
      inscription_number:
        $ref: '#/definitions/rpc.InscriptionNumber'
      inscription_output_value:
        type: integer
      ordinal_block_height:
        type: integer
      ordinal_number:
        type: integer
      ordinal_offset:
        type: integer
      satpoint_post_inscription:
        type: string
      transfers_pre_inscription:
        type: integer
      tx_index:
        type: integer
    type: object
  rpc.OrdHookInscriptionTransferred:
    properties:
      destination:
        $ref: '#/definitions/rpc.OrdHookDestination'
      inscription_id:
        type: string
      post_transfer_output_value:
        type: integer
      satpoint_post_transfer:
        type: string
      satpoint_pre_transfer:
        type: string
      tx_index:
        type: integer
    type: object
  rpc.OrdHookOrdinalOperation:
    properties:
      inscription_revealed:
        $ref: '#/definitions/rpc.OrdHookInscriptionRevealed'
      inscription_transferred:
        $ref: '#/definitions/rpc.OrdHookInscriptionTransferred'
    type: object
  rpc.OrdHookPredicate:
    properties:
      operation:
        type: string
      scope:
        type: string
    type: object
  rpc.OrdHookTransaction:
    properties:
      metadata:
        $ref: '#/definitions/rpc.OrdHookTransactionMetadata'
      operations:
        items: {}
        type: array
      transaction_identifier:
        $ref: '#/definitions/rpc.OrdHookBlockIdentifier'
    type: object
  rpc.OrdHookTransactionMetadata:
    properties:
      ordinal_operations:
        items:
          $ref: '#/definitions/rpc.OrdHookOrdinalOperation'
        type: array
      proof: {}
    type: object
...
info:
  contact: {}
  description: Mining Protocol for Inscription Miners Based on the BTC Ordinals Protocol
  termsOfService: https://github.com/SatMine/
  title: BTC Ordinals Mining Protocol
  version: "1.01"
paths:
...
  /mrc20/getrecords:
    get:
      consumes:
      - application/json
      description: Retrieves a list of records based on address, record type, page
        number, page size, and order
      parameters:
      - description: Address identifier
        in: query
        name: address
        required: true
        type: string
      - description: Record type identifier
        in: query
        name: rectype
        required: true
        type: string
      - description: Page number for pagination
        in: query
        name: pageNum
        required: true
        type: integer
      - description: Number of records per page
        in: query
        name: pageSize
        required: true
        type: integer
      - description: Order of record retrieval (true for ascending, false for descending)
        in: query
        name: ascend
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of records
          schema:
            items:
              type: string
            type: array
        "400":
          description: Error message if retrieval fails
          schema:
            type: string
      summary: Retrieve records from the database
      tags:
      - record
  /mrc20/hookevents:
    post:
      consumes:
      - application/json
      description: Parses and processes events related to OrdHook.
      parameters:
      - description: Event Payload
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/rpc.OrdHookEvent'
      produces:
      - application/json
      responses:
        "200":
          description: A message confirming successful processing
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error message in case of failure to process the event
          schema:
            additionalProperties: true
            type: object
      summary: Process OrdHook events
      tags:
      - OrdHook
...
  /mrc20/postrecord:
    post:
      consumes:
      - application/json
      description: Writes a new record with an incrementing index to the database
      parameters:
      - description: Address identifier
        in: query
        name: address
        required: true
        type: string
      - description: Record type identifier
        in: query
        name: rectype
        required: true
        type: string
      - description: Message content to be stored
        in: body
        name: message
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message with code and message text
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error message with code and message text
          schema:
            additionalProperties: true
            type: object
      summary: Write a record to the database
      tags:
      - record
...
swagger: "2.0"